Moving to the first level directory
